+++ av1/common/x86/convolve_avx2.c
@@ -158,8 +158,8 @@ void av1_convolve_y_sr_avx2(const uint8_t *src, int src_stride, uint8_t *dst,
         } else {
           __m128i *const p_0 = (__m128i *)&dst[i * dst_stride + j];
           __m128i *const p_1 = (__m128i *)&dst[i * dst_stride + j + dst_stride];
-          *(uint16_t *)p_0 = _mm_cvtsi128_si32(res_0);
-          *(uint16_t *)p_1 = _mm_cvtsi128_si32(res_1);
+          *(uint16_t *)p_0 = (uint16_t)_mm_cvtsi128_si32(res_0);
+          *(uint16_t *)p_1 = (uint16_t)_mm_cvtsi128_si32(res_1);
         }
       }
 
@@ -240,8 +240,8 @@ void av1_convolve_x_sr_avx2(const uint8_t *src, int src_stride, uint8_t *dst,
       } else {
         __m128i *const p_0 = (__m128i *)&dst[i * dst_stride];
         __m128i *const p_1 = (__m128i *)&dst[i * dst_stride + dst_stride];
-        *(uint16_t *)p_0 = _mm_cvtsi128_si32(res_0);
-        *(uint16_t *)p_1 = _mm_cvtsi128_si32(res_1);
+        *(uint16_t *)p_0 = (uint16_t)_mm_cvtsi128_si32(res_0);
+        *(uint16_t *)p_1 = (uint16_t)_mm_cvtsi128_si32(res_1);
       }
     }
   } else {
